+++ microlens-pro/CHANGELOG.md
@@ -1,3 +1,8 @@
-# 0.1.0.0
+# 0.2.0
 
-[#105](https://github.com/stevenfontanella/microlens/issues/105) microlens-pro initial release, with `Prism`s and `Iso`s.
+* New major release.
+* [#177](https://github.com/stevenfontanella/microlens/issues/177) Moves `Lens.Micro.ProCompat` into `Lens.Micro.Pro`.
+
+# 0.1.0
+
+* [#105](https://github.com/stevenfontanella/microlens/issues/105) microlens-pro initial release, with `Prism`s and `Iso`s.